eval_cma: Use dma_cma if dma_contiguous_default_area is not given
The cma evaluation module, eval_cma, is using dma_contiguous_default as
cma to be evaluated.  As a result, it cannot work when
dma_contiguous_default is not defined well.  This commit adds a fallback
for the case, which is using thp_cma as a target cma to be evaluated.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>