rcutorture/kvm.sh: Simplify logging
'kvm.sh' logs messages by printing the message to 'stdout' then writing it into
the log file.  Generation of the message occurs twice once for 'stdout', and
once for log file.  It is redundant.  Moreover, lots of the messages are
containing 'date' output.  Because the 'date' calculation also be called twice,
once for stdout print, once for log file write, the date information in stdout
and log file can be different.  It could confuse some sensitive mind people or
automation scripts.

This commit simplifies the logging by using 'tee' and fixes some trivial nits.
Such nits include messages written in log file but not printed in 'stdout', and
log file paths which is using '//' as directory seperator.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>