rcutorture/kvm.sh: Support execution from any directory
'kvm.sh' for rcutorture is restricting user to execute it from top of
linux source tree.  It is just a subtle restriction, but users using it
for the first time could forget the restriction and be confused.
Moreover, it makes commands a little longer and the long commands could
frustrate debugging people.  This commit let users to call the script
from any location.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>