rcutorture/kvm-build.sh: Skip build directory check
Check of existance and write permisison for build directory is duplicated in
'kvm-test-1-run.sh' an 'kvm-build.sh'.  Because the 'kvm-build.sh' is called
from 'kvm-test-1-run.sh' only ('kvm-build.sh' is self executable, but its not
independently executable because it depends on variables that is defined from
caller.), this commit removes the duplicated check in 'kvm-build.sh' script.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>