rcutorture/kvm-build.sh: Skip build directory check
Check of existence and write permisison for build directory exists in
'kvm-test-1-run.sh' an 'kvm-build.sh'.  Because the 'kvm-build.sh' is
dependent to 'kvm-test-1-run.sh' ('kvm-build.sh' uses variables that
defined from its caller.), it is an unnecessary duplicate.  This commit
removes, thus, the duplicated check in 'kvm-build.sh' script.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>