rcutorture/configinit: Fix build directory error message
'configinit.sh' checks format of optional argument for build directory
and print error message if the format is not valid.  However, the error
message is broken so that it just says user entered empty string as
build directory even though user entered wrong, but non-empty string as
the argument.  This commit fixes the message to show what user entered
for the argument.

Fixes: c87b9c601ac8 ("rcutorture: Add KVM-based test framework")

Signed-off-by: SeongJae Park <sj38.park@gmail.com>