rcutorture/configinit: Fix wrong build directory message
'configinit.sh' checks format of optional argument for build directory and
print error message if the format is not valid.  However, the error message is
broken so that it just says user entered empty string as build directory.  This
commit fixes the message to show what user entered for build directory optional
argument.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>